taxi_service/requirements.txt
asgiref==3.3.4
Babel==2.9.0
Django==3.2
django-flatpickr==1.0.1
django-phonenumber-field==5.0.0
phonenumbers==8.12.21
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
